+++ src/tcwrapper.cpp
@@ -349,7 +349,7 @@ public:
     }
     bool VisitTypedefDecl(TypedefDecl *TD) {
         bool isCanonical = (TD == TD->getCanonicalDecl());
-#if LLVM_VERSION >= 36 && defined(_WIN32)
+#if defined(_WIN32)
         // Starting with LLVM 3.6, clang initializes "size_t" as an implicit declaration
         // when being compatible with MSVC.
         if (!isCanonical && TD->getName().str() == "size_t") {
@@ -404,13 +404,8 @@ public:
         // foo() and not void foo(void)) we don't support old-style C parameter lists, we
         // just treat them as empty
         if (proto) {
-#if LLVM_VERSION >= 35
             for (size_t i = 0; i < proto->getNumParams(); i++) {
                 QualType PT = proto->getParamType(i);
-#else
-            for (size_t i = 0; i < proto->getNumArgs(); i++) {
-                QualType PT = proto->getArgType(i);
-#endif
                 Obj pt;
                 if (!GetType(PT, &pt)) {
                     valid = false;  // keep going with attempting to parse type to make
@@ -606,13 +601,7 @@ public:
     virtual void HandleCXXStaticMemberVarInstantiation(VarDecl *D) {
         CG->HandleCXXStaticMemberVarInstantiation(D);
     }
-#if LLVM_VERSION >= 37
     virtual void HandleVTable(CXXRecordDecl *RD) { CG->HandleVTable(RD); }
-#else
-    virtual void HandleVTable(CXXRecordDecl *RD, bool DefinitionRequired) {
-        CG->HandleVTable(RD, DefinitionRequired);
-    }
-#endif
     virtual ASTMutationListener *GetASTMutationListener() {
         return CG->GetASTMutationListener();
     }
@@ -660,37 +649,17 @@ public:
             : Name(Name_), Status(Status_), Buffer(Buffer_) {}
     virtual ~LuaProvidedFile() override {}
     virtual llvm::ErrorOr<clang::vfs::Status> status() override { return Status; }
-#if LLVM_VERSION >= 36
     virtual llvm::ErrorOr<std::unique_ptr<llvm::MemoryBuffer>> getBuffer(
             const Twine &Name, int64_t FileSize, bool RequiresNullTerminator,
             bool IsVolatile) override {
         return llvm::MemoryBuffer::getMemBuffer(Buffer, "", RequiresNullTerminator);
     }
-#else
-    virtual std::error_code getBuffer(const Twine &Name,
-                                      std::unique_ptr<llvm::MemoryBuffer> &Result,
-                                      int64_t FileSize, bool RequiresNullTerminator,
-                                      bool IsVolatile) override {
-        Result.reset(
-                llvm::MemoryBuffer::getMemBuffer(Buffer, "", RequiresNullTerminator));
-        return std::error_code();
-    }
-#endif
     virtual std::error_code close() override { return std::error_code(); }
-#if LLVM_VERSION <= 37
-    virtual void setName(StringRef Name_) override { Name = Name_; }
-#endif
 };
 #endif
 
 #if LLVM_VERSION < 50
-static llvm::sys::TimeValue ZeroTime() {
-#if LLVM_VERSION >= 36
-    return llvm::sys::TimeValue::ZeroTime();
-#else
-    return llvm::sys::TimeValue::ZeroTime;
-#endif
-}
+static llvm::sys::TimeValue ZeroTime() { return llvm::sys::TimeValue::ZeroTime(); }
 #else
 static llvm::sys::TimePoint<> ZeroTime() {
     return llvm::sys::TimePoint<>(std::chrono::nanoseconds::zero());
@@ -826,12 +795,9 @@ public:
             *contents = StringRef(data, size);
             lua_pop(L, 2);  // pop contents, size
         }
-        *status = clang::vfs::Status(Path.str(),
-#if LLVM_VERSION <= 37
-                                     "",
-#endif
-                                     clang::vfs::getNextVirtualUniqueID(), ZeroTime(), 0,
-                                     0, size, filetype, llvm::sys::fs::all_all);
+        *status = clang::vfs::Status(Path.str(), clang::vfs::getNextVirtualUniqueID(),
+                                     ZeroTime(), 0, 0, size, filetype,
+                                     llvm::sys::fs::all_all);
         lua_pop(L, 2);  // pop table, kind
         return true;
     }
@@ -849,20 +815,6 @@ public:
         }
         return llvm::errc::no_such_file_or_directory;
     }
-#if LLVM_VERSION <= 35
-    virtual std::error_code openFileForRead(
-            const llvm::Twine &Path, std::unique_ptr<clang::vfs::File> &Result) override {
-        std::error_code ec = RFS->openFileForRead(Path, Result);
-        if (!ec || ec != llvm::errc::no_such_file_or_directory) return ec;
-        clang::vfs::Status Status;
-        StringRef Buffer;
-        if (GetFile(Path, &Status, &Buffer)) {
-            Result.reset(new LuaProvidedFile(Path.str(), Status, Buffer));
-            return std::error_code();
-        }
-        return llvm::errc::no_such_file_or_directory;
-    }
-#else
     virtual llvm::ErrorOr<std::unique_ptr<clang::vfs::File>> openFileForRead(
             const llvm::Twine &Path) override {
         llvm::ErrorOr<std::unique_ptr<clang::vfs::File>> ec = RFS->openFileForRead(Path);
@@ -875,7 +827,6 @@ public:
         }
         return llvm::errc::no_such_file_or_directory;
     }
-#endif
     virtual clang::vfs::directory_iterator dir_begin(const llvm::Twine &Dir,
                                                      std::error_code &EC) override {
         printf("BUGBUG: unexpected call to directory iterator in C header include. "
@@ -887,14 +838,12 @@ public:
                                                                   // until this changes.
         return RFS->dir_begin(Dir, EC);
     }
-#if LLVM_VERSION >= 38
     llvm::ErrorOr<std::string> getCurrentWorkingDirectory() const override {
         return std::string("cwd");
     }
     std::error_code setCurrentWorkingDirectory(const Twine &Path) override {
         return std::error_code();
     }
-#endif
 };
 #endif
 
@@ -924,9 +873,7 @@ void InitHeaderSearchFlags(std::string const &TripleStr, HeaderSearchOptions &HS
     llvm::opt::ArgStringList IncludeArgs;
     TC.AddClangSystemIncludeArgs(C->getArgs(), IncludeArgs);
 
-#if LLVM_VERSION > 37
     TC.AddCudaIncludeArgs(C->getArgs(), IncludeArgs);
-#endif
 
     // organized in pairs "-<flag> <directory>"
     assert(((IncludeArgs.size() & 1) == 0) && "even number of IncludeArgs");
@@ -978,11 +925,7 @@ static void initializeclang(terra_State *T, llvm::MemoryBuffer *membuffer,
     FileManager &FileMgr = TheCompInst->getFileManager();
     TheCompInst->createSourceManager(FileMgr);
     SourceManager &SourceMgr = TheCompInst->getSourceManager();
-    TheCompInst->createPreprocessor(
-#if LLVM_VERSION >= 35
-            TU_Complete
-#endif
-    );
+    TheCompInst->createPreprocessor(TU_Complete);
     TheCompInst->createASTContext();
 
     // Set the main file handled by the source manager to the input file.
@@ -991,11 +934,7 @@ static void initializeclang(terra_State *T, llvm::MemoryBuffer *membuffer,
     TheCompInst->getDiagnosticClient().BeginSourceFile(TheCompInst->getLangOpts(),
                                                        &TheCompInst->getPreprocessor());
     Preprocessor &PP = TheCompInst->getPreprocessor();
-#if LLVM_VERSION <= 37
-    PP.getBuiltinInfo().InitializeBuiltins(PP.getIdentifierTable(), PP.getLangOpts());
-#else
     PP.getBuiltinInfo().initializeBuiltins(PP.getIdentifierTable(), PP.getLangOpts());
-#endif
 }
 
 static void AddMacro(terra_State *T, Preprocessor &PP, const IdentifierInfo *II,
@@ -1079,38 +1018,23 @@ static int dofile(terra_State *T, TerraTarget *TT, const char *code,
     // managing the various objects needed to run the compiler.
     CompilerInstance TheCompInst;
 
-#if LLVM_VERSION >= 36
     llvm::MemoryBuffer *membuffer =
             llvm::MemoryBuffer::getMemBuffer(code, "<buffer>").release();
-#else
-    llvm::MemoryBuffer *membuffer = llvm::MemoryBuffer::getMemBuffer(code, "<buffer>");
-#endif
     TheCompInst.getHeaderSearchOpts().ResourceDir = "$CLANG_RESOURCE$";
     InitHeaderSearchFlags(TT->Triple, TheCompInst.getHeaderSearchOpts());
     initializeclang(T, membuffer, args, &TheCompInst);
 
-#if LLVM_VERSION <= 36
-    CodeGenerator *codegen = CreateLLVMCodeGen(TheCompInst.getDiagnostics(), "mymodule",
-                                               TheCompInst.getCodeGenOpts(),
-                                               TheCompInst.getTargetOpts(), *TT->ctx);
-#else
     CodeGenerator *codegen = CreateLLVMCodeGen(
             TheCompInst.getDiagnostics(), "mymodule", TheCompInst.getHeaderSearchOpts(),
             TheCompInst.getPreprocessorOpts(), TheCompInst.getCodeGenOpts(), *TT->ctx);
-#endif
 
     std::stringstream ss;
     ss << "__makeeverythinginclanglive_";
     ss << TT->next_unused_id++;
     std::string livenessfunction = ss.str();
 
-#if LLVM_VERSION < 37
-    // CodeGenProxy codegenproxy(codegen,result,livenessfunction);
-    TheCompInst.setASTConsumer(new CodeGenProxy(codegen, result, TT, livenessfunction));
-#else
     TheCompInst.setASTConsumer(std::unique_ptr<ASTConsumer>(
             new CodeGenProxy(codegen, result, TT, livenessfunction)));
-#endif
 
     TheCompInst.createSema(clang::TU_Complete, NULL);
 
@@ -1128,11 +1052,7 @@ static int dofile(terra_State *T, TerraTarget *TT, const char *code,
                                       end = PP.macro_end(false);
          it != end; ++it) {
         const IdentifierInfo *II = it->first;
-#if LLVM_VERSION <= 36
-        MacroDirective *MD = it->second;
-#else
         MacroDirective *MD = it->second.getLatest();
-#endif
         AddMacro(T, PP, II, MD, &macros);
     }